chore: inline comparison benchmark support
