+++ benches/compare/amortized_input.rs
@@ -1,10 +1,5 @@
 use std::hint::black_box;
 
-#[path = "support/input.rs"]
-mod input;
-
-use input::{Input, length};
-
 fn main() {
     divan::main();
 }
@@ -29,4 +24,16 @@ mod benches {
                 assert_eq!(black_box(len), 13);
             });
     }
+}
+
+#[salsa::input]
+pub struct Input {
+    #[returns(ref)]
+    pub text: String,
+}
+
+#[salsa::tracked(returns(copy))]
+#[inline(never)]
+pub fn length(db: &dyn salsa::Database, input: Input) -> usize {
+    input.text(db).len()
 }
