+++ benches/compare/new_supertype_input.rs
@@ -1,16 +1,5 @@
 use std::hint::black_box;
 
-#[path = "support/input.rs"]
-mod input;
-#[path = "support/interned.rs"]
-mod interned;
-#[path = "support/supertype.rs"]
-mod supertype;
-
-use input::Input;
-use interned::InternedInput;
-use supertype::{SupertypeInput, either_length};
-
 fn main() {
     divan::main();
 }
@@ -54,5 +43,44 @@ mod benches {
                 let len = either_length(black_box(db), black_box(interned_input));
                 assert_eq!(black_box(len), 13);
             });
+    }
+}
+
+#[salsa::input]
+pub struct Input {
+    #[returns(ref)]
+    pub text: String,
+}
+
+#[salsa::tracked(returns(copy))]
+#[inline(never)]
+pub fn length(db: &dyn salsa::Database, input: Input) -> usize {
+    input.text(db).len()
+}
+
+#[salsa::interned]
+pub struct InternedInput<'db> {
+    #[returns(ref)]
+    pub text: String,
+}
+
+#[salsa::tracked(returns(copy))]
+#[inline(never)]
+pub fn interned_length<'db>(db: &'db dyn salsa::Database, input: InternedInput<'db>) -> usize {
+    input.text(db).len()
+}
+
+#[derive(Clone, Copy, PartialEq, Eq, Hash, salsa::Supertype)]
+enum SupertypeInput<'db> {
+    InternedInput(InternedInput<'db>),
+    Input(Input),
+}
+
+#[salsa::tracked(returns(copy))]
+#[inline(never)]
+pub fn either_length<'db>(db: &'db dyn salsa::Database, input: SupertypeInput<'db>) -> usize {
+    match input {
+        SupertypeInput::InternedInput(input) => interned_length(db, input),
+        SupertypeInput::Input(input) => length(db, input),
     }
 }
