--- benches/compare/support/input.rs
@@ -1,11 +1,0 @@
-#[salsa::input]
-pub struct Input {
-    #[returns(ref)]
-    pub text: String,
-}
-
-#[salsa::tracked(returns(copy))]
-#[inline(never)]
-pub fn length(db: &dyn salsa::Database, input: Input) -> usize {
-    input.text(db).len()
-}
